Added error handling for a weird bug

+++ professor.py
@@ -408,16 +408,17 @@ async def on_message(message):
 
     # Check if we are in dm
     guildMessage = isinstance(message.channel, discord.abc.GuildChannel)
-    if guildMessage \
-            and message.channel == eventsDict[hash(message.guild)].channel \
-            and (message.author != professor.user or str(message.type) == "MessageType.pins_add") \
-            and eventsDict[hash(message.guild)].scheduling == 0:
-        try:
-            await message.delete()
-        except discord.errors.NotFound:
-            pass
-        except discord.errors.Forbidden:
-            pass
+    if hash(message.guild) in eventsDict.keys():
+        if guildMessage \
+                and message.channel == eventsDict[hash(message.guild)].channel \
+                and (message.author != professor.user or str(message.type) == "MessageType.pins_add") \
+                and eventsDict[hash(message.guild)].scheduling == 0:
+            try:
+                await message.delete()
+            except discord.errors.NotFound:
+                pass
+            except discord.errors.Forbidden:
+                pass
 
 
 @professor.event
